Backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
sqlalchemy
langchain
langchain-google-genai
langchain-core